# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DAMON Bot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStatus {
  online: boolean;
  /** @nullable */
  uid: string | null;
  /** @nullable */
  name: string | null;
  uptimeSeconds: number;
  /** @nullable */
  startedAt: string | null;
  /** @nullable */
  connectionType: string | null;
  adminCount: number;
  cookieLoaded: boolean;
}

export interface Admin {
  uid: string;
  name: string;
  addedAt: string;
}

export interface AdminInput {
  uid: string;
  name: string;
}

export interface CookieInfo {
  loaded: boolean;
  /** @nullable */
  preview: string | null;
  /** @nullable */
  updatedAt: string | null;
}

export interface CookieInput {
  /** c3c JSON cookie string (array of cookie objects) */
  cookie: string;
}

export interface LogEntry {
  id: string;
  level: string;
  tag: string;
  message: string;
  timestamp: string;
}

export interface ActionResult {
  success: boolean;
  message: string;
}

export type GetLogsParams = {
limit?: number;
};