# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DAMON Bot API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admins:
    get:
      operationId: listAdmins
      tags: [admins]
      summary: List all bot admins
      responses:
        "200":
          description: List of admins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admin"
    post:
      operationId: addAdmin
      tags: [admins]
      summary: Add a bot admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInput"
      responses:
        "201":
          description: Admin added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "409":
          description: Already an admin
  /admins/{uid}:
    delete:
      operationId: removeAdmin
      tags: [admins]
      summary: Remove a bot admin
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Admin removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "404":
          description: Admin not found
...
components:
  schemas:
...
    Admin:
      type: object
      properties:
        uid:
          type: string
        name:
          type: string
        addedAt:
          type: string
      required:
        - uid
        - name
        - addedAt
    AdminInput:
      type: object
      properties:
        uid:
          type: string
        name:
          type: string
      required:
        - uid
        - name
...
    ActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message